initial refactor done
