+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/system/events/CesEvent.java
@@ -49,7 +49,8 @@ public class CesEvent extends AbstractSerializableHashable
     public static final long UNDEFINED = -1;
 
     public static final long CLASS_ID = 0xe250a9fbdcc4b1baL;
-    public static final int CLASS_VERSION = 1;
+    private static final int CLASS_VERSION = 1;
+    private static final int CONSENSUS_DATA_CLASS_VERSION = 2;
 
     /** the pre-consensus event */
     private PlatformEvent platformEvent;
@@ -95,7 +96,7 @@ public class CesEvent extends AbstractSerializableHashable
         // some fields used to be part of the stream but are no longer used
         // in order to maintain compatibility with older versions of the stream, we write a constant in their place
 
-        out.writeInt(ConsensusData.CLASS_VERSION);
+        out.writeInt(CONSENSUS_DATA_CLASS_VERSION);
         out.writeLong(UNDEFINED); // ConsensusData.generation
         out.writeLong(UNDEFINED); // ConsensusData.roundCreated
         out.writeBoolean(false); // ConsensusData.stale
